+++ tests/unit/structures/test_workflow.py
@@ -423,6 +423,7 @@ class TestWorkflow:
 
     def test_run_topology_2_imperative_parents(self):
         taska = PromptTask("testa", id="taska")
+        workflow = Workflow(prompt_driver=MockPromptDriver(), tasks=[taska])
         taskb = PromptTask("testb", id="taskb")
         taskc = PromptTask("testc", id="taskc")
         taskd = PromptTask("testd", id="taskd")
@@ -431,7 +432,6 @@ class TestWorkflow:
         taskc.add_parent("taska")
         taskd.add_parents([taska, taskb, taskc])
         taske.add_parents(["taska", taskd, "taskc"])
-        workflow = Workflow(prompt_driver=MockPromptDriver(), tasks=[taska, taskb, taskc, taskd, taske])
 
         workflow.run()
 
@@ -439,6 +439,7 @@ class TestWorkflow:
 
     def test_run_topology_2_imperative_children(self):
         taska = PromptTask("testa", id="taska")
+        workflow = Workflow(prompt_driver=MockPromptDriver(), tasks=[taska])
         taskb = PromptTask("testb", id="taskb")
         taskc = PromptTask("testc", id="taskc")
         taskd = PromptTask("testd", id="taskd")
@@ -447,7 +448,6 @@ class TestWorkflow:
         taskb.add_child(taskd)
         taskc.add_children([taskd, taske])
         taskd.add_child(taske)
-        workflow = Workflow(prompt_driver=MockPromptDriver(), tasks=[taska, taskb, taskc, taskd, taske])
 
         workflow.run()